Refactored location of SAMHSA utility classes.
