+++ apps/bfd-model/bfd-model-rda/src/main/java/gov/cms/bfd/model/rda/samhsa/FissTag.java
@@ -32,7 +32,7 @@ public class FissTag {
   /** the details field. */
   @JdbcTypeCode(SqlTypes.JSON)
   @Column(name = "details", nullable = true)
-  private List<TagDetails> details;
+  private List details;
 
   /** The associated claim. */
   @Id
