+++ apps/bfd-model/bfd-model-rda/src/main/java/gov/cms/bfd/model/rda/samhsa/McsTag.java
@@ -32,7 +32,7 @@ public class McsTag {
   /** The tag details. */
   @JdbcTypeCode(SqlTypes.JSON)
   @Column(name = "details", nullable = true)
-  private List<TagDetails> details;
+  private List details;
 
   /** The associated claim. */
   @Id
