+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/SamhsaUtil.java
@@ -11,9 +11,9 @@ import gov.cms.bfd.model.rda.entities.RdaMcsDetail;
 import gov.cms.bfd.model.rda.entities.RdaMcsDiagnosisCode;
 import gov.cms.bfd.model.rda.samhsa.FissTag;
 import gov.cms.bfd.model.rda.samhsa.McsTag;
-import gov.cms.bfd.model.rda.samhsa.SamhsaEntry;
-import gov.cms.bfd.model.rda.samhsa.TagCode;
-import gov.cms.bfd.model.rda.samhsa.TagDetails;
+import gov.cms.bfd.pipeline.sharedutils.model.SamhsaEntry;
+import gov.cms.bfd.pipeline.sharedutils.model.TagCode;
+import gov.cms.bfd.pipeline.sharedutils.model.TagDetails;
 import jakarta.persistence.EntityManager;
 import java.io.IOException;
 import java.io.InputStream;
